feat(airlines): move constants to constants.js
