+++ src/constants.js
@@ -1,0 +1,2 @@
+export const HOTEL_SEGMENT_ID = 'hotels';
+export const AIRLINE_SEGMENT_ID = 'airlines';
