+++ src/index.js
@@ -7,8 +7,7 @@ import type { WTHotelIndexInterface } from './interfaces/hotel-interfaces';
 import type { WTAirlineIndexInterface } from './interfaces/airline-interfaces';
 import { AbstractDataModel, AirlineDataModel, HotelDataModel } from './data-model/index';
 import OffChainDataClient from './off-chain-data-client';
-import { HOTEL_SEGMENT_ID } from '../test/utils/data-hotel-model-definition';
-import { AIRLINE_SEGMENT_ID } from '../test/utils/data-airline-model-definition';
+import { AIRLINE_SEGMENT_ID, HOTEL_SEGMENT_ID } from './constants';
 
 import {
   WTLibsError,
